Удаляет моковые данные, блокирует кнопку создания нового события при загрузке
