--- src/mock/filter.js
@@ -1,8 +1,0 @@
-import { filter } from '../utils.js';
-
-export const generateFilter = (wayPoints) => Object.entries(filter).map(
-  ([filterName, filterWayPoints]) => ({
-    name: filterName,
-    count: filterWayPoints(wayPoints).length,
-  }),
-);
