+++ src/presenter/filter-presenter.js
@@ -1,7 +1,7 @@
 import { render, replace, remove } from '../framework/render.js';
 import FilterView from '../view/filter-view.js';
 import { filter } from '../utils.js';
-import { FilterType, UpdateType } from '../mock/const.js';
+import { FilterType, UpdateType } from '../const.js';
 
 export default class FilterPresenter {
   #filterContainer = null;
