+++ src/view/sort-view.js
@@ -1,5 +1,5 @@
 import AbstractView from '../framework/view/abstract-view';
-import { SortType } from '../mock/const.js';
+import { SortType } from '../const.js';
 
 const createSortTemplate = (currentSortType) =>
   (`<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
